log an error if socket read returns 0 and reset the socket (#67)
In a couple use cases, I observed that the UDP socket bound to the mDNS port 5353 disappeared after some time, and the daemon no longer receives packets.

It is not clear when or how the socket was closed or shutdown. But an error message as below indicates that sock.read probably returns 0. Hence adding a log to catch such cases to help future debugging.

2022-09-24T21:56:05.612435Z ERROR ThreadId(09) mdns_sd::service_daemon: Invalid incoming message: read_name: from_utf8: invalid utf-8 sequence of 1 bytes from index 0

Fix: create a new socket to replace the closed one. I don't know how to reproduce the issue, so do not have a test case yet.