bugfix: only remove a service instance if all its SRV are gone (#350)
When SRV refreshes with a different host (see in the log), we expire the old SRV record (per RFC, expire in 1 second), and add a new SRV record under the same service instance name (i.e. SRV name keeps the same).

The problem is: when the old SRV expires at the end of 1 second, we are removing not only the old SRV, but the list of all SRV records of the service instance, including the new SRV record. This triggered the service instance itself got removed.

Hence, the fix is that we only remove the old SRV record and keep the rest of SRV list for a given name.