refactoring: define our own interface struct to allow multiple addresses (#380)
This is triggered by #377 as currently we would skip 2nd IP address on an interface due to join_multicast_group error on the same interface. Strictly speaking, this is wrong and it can cause confusion in debugging as addr 1 is in ZeroConf.my_intfs but addr 2 (on the same interface) is not.

Also, currently we directly use Interface from if-addrs crate, but it is per address, not truly per interface. I took this opportunity to define a new interface struct for us MyIntf that represents a network interface that supports multiple addresses.

The diff is very extensive but conceptually really just Interface is replaced by MyIntf.